# communication-hub-ui/src/app/api/models/communication-message-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AttachmentDto } from '../models/attachment-dto';
export interface CommunicationMessageDto {
  attachments?: (Array<AttachmentDto> | null);
  communicationId?: string;
  direction?: (string | null);
  isRead?: (boolean | null);
  messageBody?: (string | null);
  mode?: (string | null);
  notes?: (string | null);
  status?: (string | null);
  timestamp?: (string | null);
}