# communication-hub-ui/src/app/api/models/communication-thread-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CommunicationMessageDto } from '../models/communication-message-dto';
export interface CommunicationThreadDto {
  claimId?: (number | string);
  claimNumber?: (string | null);
  messages?: (Array<CommunicationMessageDto> | null);
  partyId?: (number | string);
  partyName?: (string | null);
  policyNumber?: (string | null);
}